Add PlayerGuild and PlayerGuildItems models
- Populate player guild info/bank from addon data

+++ src/Wowthing.Backend/Jobs/JobBase.cs
@@ -71,7 +71,7 @@ namespace Wowthing.Backend.Jobs
             return LogContext.PushProperty("Task", $"{userId} {jobName}");
         }
 
-        protected IDisposable UserLog(int userId) => UserLog(userId.ToString());
+        protected IDisposable UserLog(long userId) => UserLog(userId.ToString());
 
         protected static Uri GenerateUri(WowRegion region, ApiNamespace lamespace, string path)
         {
